requirements.txt
numpy
matplotlib
h5py
scipy
Pillow  # PIL
